style, feat: overhauled styling of catalog and nav buttons in gallery to match homepage. Added sorting
