+++ app/catalog/catalog.styles.css
@@ -3,7 +3,7 @@
 
 /* Extend the base styles from homepage - using the same variables */
 :root {
-  /* Main color palette */
+  /* Main color palette - original values */
   --color-bg: #f7f3ee;
   --color-text: #453c33;
   --color-text-secondary: #635a50;
@@ -12,12 +12,27 @@
   --color-border-light: #d7c9b8;
   --color-bg-subtle: #ece5dc;
 
-  /* Accent colors */
+  /* Accent color for key interactions - original brown */
   --color-accent: #7c6b4e;
   --color-accent-light: rgba(124, 107, 78, 0.3);
   --color-accent-hover: #9f896b;
+
+  /* Secondary accent for visual variety */
   --color-accent-secondary: #8a6246;
+
+  /* Highlight color for important elements */
   --color-highlight: #9c6b4e;
+}
+
+body {
+  margin: 0;
+  padding: 0;
+  background-color: var(--color-bg);
+  position: relative;
+  overflow-x: hidden;
+  color: var(--color-text);
+  font-family: "IBM Plex Mono", monospace;
+  line-height: 1.5;
 }
 
 /* CircuitBoard background - exactly the same as homepage */
@@ -389,11 +404,16 @@
   transition: all 0.3s ease;
   border-radius: 4px;
   position: relative;
+  transition: all 0.25s ease;
+  overflow: hidden;
 }
 
 .incident-item:hover {
   background-color: var(--color-bg-subtle);
   border-left-color: var(--color-accent-hover);
+  transform: translateY(-3px);
+  box-shadow: 0 6px 12px rgba(90, 83, 72, 0.15);
+  border-left-width: 5px;
   transform: translateY(-2px);
   box-shadow: 0 4px 8px rgba(90, 83, 72, 0.1);
 }
@@ -403,6 +423,8 @@
   font-size: 1.25rem; /* Reduced from 1.5rem */
   margin-bottom: 0.3rem; /* Reduced from 0.5rem */
   line-height: 1;
+  display: flex;
+  align-items: center;
 }
 
 .incident-name {
@@ -648,4 +670,244 @@ body {
   color: var(--color-text);
   font-family: "IBM Plex Mono", monospace;
   line-height: 1.5;
+}
+
+/* Enhanced card styles with hover effects and visual markers */
+.incident-item {
+  transition: all 0.25s ease;
+  position: relative;
+  overflow: hidden;
+}
+
+.incident-item:hover {
+  transform: translateY(-3px);
+  box-shadow: 0 6px 12px rgba(90, 83, 72, 0.15);
+  border-left-width: 5px;
+}
+
+/* Add severity indicators */
+.severity-indicator {
+  position: absolute;
+  top: 0.75rem;
+  right: 0.75rem;
+  display: flex;
+  gap: 2px;
+}
+
+.severity-dot {
+  width: 6px;
+  height: 6px;
+  border-radius: 50%;
+  background-color: var(--color-border-light);
+}
+
+.severity-dot.active {
+  background-color: var(--color-accent);
+}
+
+/* High severity marker */
+.incident-item.high-severity::before {
+  content: "";
+  position: absolute;
+  top: -12px;
+  right: -12px;
+  width: 40px;
+  height: 40px;
+  background-color: #c75146;
+  transform: rotate(45deg);
+  z-index: 1;
+}
+
+.incident-item.high-severity::after {
+  content: "!";
+  position: absolute;
+  top: 3px;
+  right: 5px;
+  color: white;
+  font-weight: bold;
+  font-size: 10px;
+  z-index: 2;
+}
+
+/* Enhance metadata display */
+.incident-year {
+  display: flex;
+  align-items: center;
+}
+
+/* Category icons */
+.category-icon {
+  margin-right: 0.5rem;
+  width: 14px;
+  height: 14px;
+  opacity: 0.7;
+}
+
+/* Add teaser for incident description */
+.incident-teaser {
+  font-size: 0.75rem;
+  color: var(--color-text-light);
+  margin-top: 0.3rem;
+  line-height: 1.3;
+  max-height: 2.6rem;
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  font-style: italic;
+}
+
+/* Add view details button that appears on hover */
+.view-details {
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
+  background-color: var(--color-accent);
+  color: white;
+  text-align: center;
+  padding: 0.4rem;
+  font-size: 0.75rem;
+  transform: translateY(100%);
+  transition: transform 0.2s ease-in-out;
+}
+
+.incident-item:hover .view-details {
+  transform: translateY(0);
+}
+
+/* Enhance filter UI */
+.filter-section {
+  margin: 1rem 0;
+}
+
+.filter-header {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-bottom: 0.5rem;
+}
+
+.filter-title {
+  font-size: 0.8rem;
+  color: var(--color-text-secondary);
+  text-transform: uppercase;
+  letter-spacing: 1px;
+}
+
+/* Category filters */
+.category-filters {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+  margin: 0.7rem 0;
+}
+
+.category-filter {
+  display: flex;
+  align-items: center;
+  gap: 0.3rem;
+  padding: 0.3rem 0.7rem;
+  background-color: var(--color-bg-subtle);
+  border: 1px solid var(--color-border-light);
+  border-radius: 3px;
+  font-size: 0.75rem;
+  cursor: pointer;
+  transition: all 0.2s ease;
+}
+
+.category-filter.active {
+  background-color: var(--color-accent-light);
+  border-color: var(--color-accent);
+}
+
+.category-filter:hover:not(.active) {
+  background-color: var(--color-bg);
+  border-color: var(--color-accent-light);
+}
+
+/* Sort controls */
+.sort-controls {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+  margin: 0.7rem 0;
+}
+
+.sort-label {
+  font-size: 0.8rem;
+  color: var(--color-text-secondary);
+}
+
+.sort-select {
+  font-family: "IBM Plex Mono", monospace;
+  font-size: 0.8rem;
+  padding: 0.3rem 0.5rem;
+  background-color: var(--color-bg-subtle);
+  border: 1px solid var(--color-border-light);
+  border-radius: 3px;
+  color: var(--color-text);
+}
+
+/* Enhanced subtitle */
+.archive-subtitle {
+  font-family: "IBM Plex Mono", monospace;
+  font-size: 1rem;
+  color: var(--color-text-secondary);
+  margin: -0.5rem 0 1.5rem;
+  font-weight: normal;
+}
+
+/* Timeline visualization */
+.timeline {
+  position: relative;
+  height: 50px;
+  margin: 1rem 0;
+  background-color: var(--color-bg-subtle);
+  border-radius: 4px;
+  overflow: hidden;
+}
+
+.timeline-marker {
+  position: absolute;
+  top: 0;
+  width: 2px;
+  height: 100%;
+  background-color: var(--color-accent);
+  transform: translateX(-50%);
+}
+
+.timeline-marker::after {
+  content: attr(data-year);
+  position: absolute;
+  bottom: -20px;
+  left: 50%;
+  transform: translateX(-50%);
+  font-size: 0.7rem;
+  color: var(--color-text-secondary);
+}
+
+.timeline-marker.major {
+  background-color: var(--color-accent-secondary);
+  width: 4px;
+}
+
+/* Loading state and animations */
+@keyframes pulse {
+  0% {
+    opacity: 0.6;
+  }
+  50% {
+    opacity: 0.3;
+  }
+  100% {
+    opacity: 0.6;
+  }
+}
+
+.loading-card {
+  height: 120px;
+  background-color: var(--color-bg-subtle);
+  border-radius: 4px;
+  animation: pulse 1.5s infinite;
 }
