+++ app/components/ui/gallery-navigation/gallery-variables.css
@@ -1,0 +1,21 @@
+:root {
+  /* Main color palette - from homepage.styles.css */
+  --color-bg: #f7f3ee;
+  --color-text: #453c33;
+  --color-text-secondary: #635a50;
+  --color-text-light: #7d7165;
+  --color-border: #c9b8a8;
+  --color-border-light: #d7c9b8;
+  --color-bg-subtle: #ece5dc;
+
+  /* Accent color for key interactions */
+  --color-accent: #7c6b4e;
+  --color-accent-light: rgba(124, 107, 78, 0.3);
+  --color-accent-hover: #9f896b;
+
+  /* Secondary accent for visual variety */
+  --color-accent-secondary: #8a6246;
+
+  /* Highlight color for important elements */
+  --color-highlight: #9c6b4e;
+}
