+++ app/components/ui/gallery-navigation/slide-transitions.css
@@ -7,12 +7,13 @@
   z-index: 1000;
   display: flex;
   align-items: center;
-  background-color: #c0c0c0; /* Classic Win95 gray */
-  border: 2px solid;
-  border-color: #dfdfdf #555 #555 #dfdfdf;
-  box-shadow: 1px 1px 0 #000;
-  padding: 5px;
-  min-height: 30px;
+  background-color: rgba(245, 242, 237, 0.96);
+  border: 1px solid var(--color-border, #c9b8a8);
+  box-shadow: 0 4px 15px rgba(73, 107, 99, 0.15), 0 1px 3px rgba(107, 92, 76, 0.2);
+  padding: 10px;
+  min-height: 35px;
+  font-family: "IBM Plex Mono", monospace;
+  border-radius: 8px;
 }
 
 /* The timeline section that shows the years */
@@ -31,12 +32,10 @@
 
 /* Year button styling */
 .year-button {
-  background-color: #c0c0c0;
-  border-width: 2px;
-  border-style: solid;
-  border-color: #dfdfdf #555 #555 #dfdfdf;
-  padding: 2px 8px;
-  font-family: "MS Sans Serif", "Tahoma", sans-serif;
+  background-color: var(--color-bg-subtle, #ece5dc);
+  border: 1px solid var(--color-border, #c9b8a8);
+  padding: 4px 8px;
+  font-family: "IBM Plex Mono", monospace;
   font-size: 14px;
   cursor: pointer;
   min-width: 50px;
@@ -45,39 +44,31 @@
   position: relative;
   user-select: none; /* Prevent text selection while clicking */
   margin: 0 5px;
+  color: var(--color-text, #453c33);
+  transition: all 0.3s ease;
+  border-radius: 4px;
+  box-shadow: 0 2px 5px rgba(73, 107, 99, 0.1);
 }
 
 .year-button:active {
-  border-color: #555 #dfdfdf #dfdfdf #555;
-  box-shadow: none;
-  transform: translate(
-    1px,
-    1px
-  ); /* Use transform instead of padding to create press effect */
+  box-shadow: inset 0 1px 3px rgba(73, 107, 99, 0.2);
+  transform: translateY(1px);
+}
+
+.year-button:hover {
+  background-color: var(--color-bg, #f7f3ee);
+  box-shadow: 0 3px 8px rgba(73, 107, 99, 0.15);
 }
 
 .year-button--active {
-  background-color: #000080; /* Win95 selection blue */
+  background-color: var(--color-accent, #7c6b4e);
   color: white;
   font-weight: bold;
   z-index: 1;
-}
-
-/* Apply a subtle glow to the active year for emphasis */
-.year-button--active::after {
-  content: "";
-  position: absolute;
-  inset: -1px;
-  border: 1px solid #ffffff55;
-  pointer-events: none;
+  box-shadow: 0 3px 8px rgba(124, 107, 78, 0.3);
 }
 
 /* Active year button highlight animation */
-.year-button--active {
-  position: relative;
-  overflow: hidden;
-}
-
 .year-button--active::before {
   content: "";
   position: absolute;
@@ -102,35 +93,64 @@
 
 /* Catalog button styling */
 .catalog-button {
-  background-color: #c0c0c0;
-  border-width: 2px;
-  border-style: solid;
-  border-color: #dfdfdf #555 #555 #dfdfdf;
-  padding: 2px 8px;
-  font-family: "MS Sans Serif", "Tahoma", sans-serif;
+  background-color: var(--color-accent, #7c6b4e);
+  color: white;
+  border: none;
+  padding: 8px 16px;
+  font-family: "IBM Plex Mono", monospace;
   font-size: 14px;
   cursor: pointer;
-  min-width: 70px;
+  min-width: 80px;
   text-align: center;
   position: relative;
   user-select: none;
-  margin-left: 10px; /* Add spacing between timeline and catalog button */
+  margin-left: 12px; /* Add spacing between timeline and catalog button */
+  transition: all 0.3s ease;
+  box-shadow: 0 2px 8px rgba(73, 107, 99, 0.2);
+  border-radius: 6px;
+  letter-spacing: 0.5px;
+  text-transform: uppercase;
+  font-weight: 600;
+}
+
+.catalog-button:hover {
+  background-color: var(--color-accent-hover, #9f896b);
+  box-shadow: 0 4px 12px rgba(73, 107, 99, 0.3);
+  transform: translateY(-1px);
 }
 
 .catalog-button:active {
-  border-color: #555 #dfdfdf #dfdfdf #555;
-  box-shadow: none;
-  transform: translate(1px, 1px);
+  box-shadow: 0 1px 3px rgba(73, 107, 99, 0.2);
+  transform: translateY(1px);
+}
+
+/* Add glow effect to match homepage buttons */
+.catalog-button::after {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: radial-gradient(circle at center, rgba(255,255,255,0.4) 0%, transparent 70%);
+  opacity: 0;
+  transition: opacity 0.3s ease;
+  border-radius: 6px;
+}
+
+.catalog-button:hover::after {
+  opacity: 0.5;
 }
 
 /* Placeholder for when no years are available */
 .year-placeholder {
-  font-family: "MS Sans Serif", "Tahoma", sans-serif;
+  font-family: "IBM Plex Mono", monospace;
   font-size: 14px;
-  color: #777;
+  color: var(--color-text-light, #7d7165);
   text-align: center;
-  padding: 2px 5px;
+  padding: 4px 5px;
   width: 100%;
+  font-style: italic;
 }
 
 /* Responsive adjustments */
@@ -142,16 +162,22 @@
   .year-button {
     min-width: 40px;
     flex: 0 0 40px;
-    padding: 1px 6px;
+    padding: 3px 6px;
   }
 
   .catalog-button {
-    padding: 1px 6px;
-    min-width: 60px;
+    padding: 6px 12px;
+    min-width: 70px;
+    font-size: 13px;
   }
 }
 
 @media (max-width: 480px) {
+  .timeline-container {
+    padding: 8px;
+    border-radius: 6px;
+  }
+
   .timeline-inner-container {
     width: 210px;
   }
@@ -159,14 +185,17 @@
   .year-button {
     min-width: 35px;
     flex: 0 0 35px;
-    padding: 1px 5px;
+    padding: 2px 5px;
     font-size: 12px;
     margin: 0 3px;
+    border-radius: 3px;
   }
 
   .catalog-button {
-    padding: 1px 5px;
-    min-width: 50px;
+    padding: 5px 10px;
+    min-width: 60px;
     font-size: 12px;
+    margin-left: 8px;
+    border-radius: 4px;
   }
 }
